implement executor. everything working non-ssr
