+++ src/query_client.rs
@@ -8,6 +8,11 @@ use std::{
     rc::Rc,
 };
 
+/// Retrieves a Query Client from the current scope.
+pub fn use_query_client(cx: Scope) -> QueryClient {
+    use_context::<QueryClient>(cx).expect("Query Client Missing.")
+}
+
 /// The Cache Client to store query data.
 /// Exposes utility functions to manage queries.
 #[derive(Clone)]
@@ -47,7 +52,7 @@ impl QueryClient {
         K: Hash + Eq + PartialEq + Clone + 'static,
         V: Clone + 'static,
     {
-        self.use_cache(|cache: &HashMap<K, QueryState<K, V>>| {
+        self.use_cache_option(|cache: &HashMap<K, QueryState<K, V>>| {
             cache.get(key).map(|state| state.invalidate())
         })
         .is_some()
@@ -83,7 +88,7 @@ impl QueryClient {
         None
     }
 
-    fn use_cache<K, V, R, F>(&self, func: F) -> Option<R>
+    fn use_cache_option<K, V, R, F>(&self, func: F) -> Option<R>
     where
         K: Clone + 'static,
         V: Clone + 'static,
@@ -98,4 +103,30 @@ impl QueryClient {
         }
         None
     }
+}
+
+pub(crate) fn use_cache<K, V, R>(
+    cx: Scope,
+    func: impl FnOnce((Scope, &mut HashMap<K, QueryState<K, V>>)) -> R + 'static,
+) -> R
+where
+    K: 'static,
+    V: 'static,
+{
+    let client = use_query_client(cx);
+    let mut cache = client.cache.borrow_mut();
+    let entry = cache.entry(TypeId::of::<K>());
+
+    let cache = entry.or_insert_with(|| {
+        let wrapped: CacheEntry<K, V> = Rc::new(RefCell::new(HashMap::new()));
+        let boxed = Box::new(wrapped) as Box<dyn Any>;
+        boxed
+    });
+
+    let mut cache = cache
+        .downcast_ref::<CacheEntry<K, V>>()
+        .expect("Query Cache Type Mismatch.")
+        .borrow_mut();
+
+    func((client.cx, &mut cache))
 }
