+++ src/query_result.rs
@@ -1,3 +1,5 @@
+use std::rc::Rc;
+
 use crate::{
     instant::Instant,
     query_state::QueryState,
@@ -38,17 +40,14 @@ where
     pub(crate) fn new<K: Clone>(
         cx: Scope,
         state: Memo<QueryState<K, V>>,
-        resource: Resource<K, Option<V>>,
+        data: Signal<Option<V>>,
+        refetch: Rc<dyn Fn() -> ()>,
     ) -> QueryResult<V> {
-        let data = Signal::derive(cx, move || resource.read(cx).flatten());
         let is_loading = Signal::derive(cx, move || state.get().is_loading(cx).get());
         let is_stale = make_stale_signal(cx, state);
         let is_refetching = Signal::derive(cx, move || state.get().fetching.get());
         let updated_at = Signal::derive(cx, move || state.get().updated_at.get());
-        let refetch = move |_: ()| {
-            state.get().needs_refetch.set(true);
-            resource.refetch()
-        };
+        let refetch = move |_: ()| refetch();
 
         QueryResult {
             data,
@@ -65,7 +64,7 @@ impl<V: Copy> Copy for QueryResult<V> where V: 'static {}
 
 fn make_stale_signal<K: Clone, V: Clone>(cx: Scope, state: Memo<QueryState<K, V>>) -> Signal<bool> {
     let (stale, set_stale) = create_signal(cx, false);
-    create_effect(cx, move |_| {
+    create_isomorphic_effect(cx, move |_| {
         let state = state.get();
         let updated_at = state.updated_at;
         let stale_time = state.stale_time;
@@ -73,16 +72,19 @@ fn make_stale_signal<K: Clone, V: Clone>(cx: Scope, state: Memo<QueryState<K, V>
         use_timeout(cx, move || match (updated_at.get(), stale_time.get()) {
             (Some(updated_at), Some(stale_time)) => {
                 let timeout = time_until_stale(updated_at, stale_time);
-                if !timeout.is_zero() {
+                if timeout.is_zero() {
+                    set_stale.set(true);
+                    None
+                } else {
                     set_stale.set(false);
+                    set_timeout_with_handle(
+                        move || {
+                            set_stale.set(true);
+                        },
+                        timeout,
+                    )
+                    .ok()
                 }
-                set_timeout_with_handle(
-                    move || {
-                        set_stale.set(true);
-                    },
-                    timeout,
-                )
-                .ok()
             }
             _ => None,
         })
